// src/app/opmobilitybackend/models/user-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserRoleResponseDto } from '../models/user-role-response-dto';
export interface UserResponseDto {
  createdAt: string;
  email: string;
  firstName: string;
  id: number;
  isDeactivated: boolean;
  lastName: string;
  phone: string;
  profilePictureUrl: string;
  role: UserRoleResponseDto;
  updatedAt: string;
}